Insert the application specific hook
During the CMD set of commands as the container is starting up there is
a hook to the application specific script to do anything needed before
the container is actually ready to do what it needs to.  Think of it as
a last second hook to change things in the container before it actually
starts.