Change REGION to BACKUP_S3_REGION
The value for the REGION that is passed into this script was to
identify the AWS_REGION for S3.  So, the name has been change to
reflect it's meaning: BACKUP_S3_REGION.  Also, this variable is now
defined in the appArchitecure json file and is passed in here because
the S3 region is NOT the same as the ASW EC2 Region.  So, it could be
different from the AWS_REGION (even though it reads from an AWS_REGION
variable.